% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tellaus.r
\name{add_metrics}
\alias{add_metrics}
\title{Add metric labels (e.g. r2, rmse, etc.) to a plot.}
\usage{
add_metrics(benchmark, comparisons, settings, plt)
}
\arguments{
\item{benchmark}{A Benchmark object.}

\item{comparisons}{Output of DGVMBenchmarks::fullSpatialComparison()}

\item{settings}{Benchmark settings, as defined in tellaus.rmd.}

\item{plt}{A ggplot object generated by \code{\link[DGVMTools:plotSpatialComparison]{DGVMTools::plotSpatialComparison()}}}
}
\value{
\link{ggplot} A plot with fixed x-axis labels.
}
\description{
Add metric labels (e.g. r2, rmse, etc.) to a plot.
}
\keyword{internal}
