% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.r
\name{aggregate_to}
\alias{aggregate_to}
\title{Aggregate the given data frame to the grid specified by the given set of
coordinates, by taking the mean of the nearest neighbours on the target grid.}
\usage{
aggregate_to(df, lons, lats)
}
\arguments{
\item{df}{A data frame.}

\item{lons}{List of longitude values on the target grid.}

\item{lats}{List of latitude values on the target grid.}
}
\value{
Returns a data frame on a the target grid.
}
\description{
Aggregate the given data frame to the grid specified by the given set of
coordinates, by taking the mean of the nearest neighbours on the target grid.
}
