% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.r
\name{aggregate_to_1deg}
\alias{aggregate_to_1deg}
\title{Aggregate the given data frame to 1-degree by taking the mean of the closest
neighbours to each gridcell on a 1-degree grid.}
\usage{
aggregate_to_1deg(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
Returns a data frame on a 1-degree grid.
}
\description{
Aggregate the given data frame to 1-degree by taking the mean of the closest
neighbours to each gridcell on a 1-degree grid.
}
