% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-auseflux.r
\name{auseflux_get_file_path}
\alias{auseflux_get_file_path}
\title{Get the path to the AusEFlux data file for the specified variable.}
\usage{
auseflux_get_file_path(var)
}
\arguments{
\item{var}{The AusEFlux variable.}
}
\description{
Get the path to the AusEFlux data file for the specified variable.
}
\keyword{internal}
