% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-auseflux.r
\name{auseflux_sanitise_var}
\alias{auseflux_sanitise_var}
\title{Attempt to interpret the specified variable as a valid AusEFlux variable, and
return the name of the corresponding AusEFlux variable. stop() will be called
if the variable is invalid.}
\usage{
auseflux_sanitise_var(var)
}
\arguments{
\item{var}{The AusEFlux variable.}
}
\description{
Attempt to interpret the specified variable as a valid AusEFlux variable, and
return the name of the corresponding AusEFlux variable. stop() will be called
if the variable is invalid.
}
\keyword{internal}
