% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-csv.r
\name{available_layers_csv}
\alias{available_layers_csv}
\title{List the available layers (columns) in a CSV file.}
\usage{
available_layers_csv(source, quant)
}
\arguments{
\item{source}{A \code{\linkS4class{Source}} containing the meta-data about
the CSV source.}

\item{quant}{A \code{\linkS4class{Quantity}} object to specify what quantity
should be opened.}
}
\value{
A character vector of layer names available for this quantity.
}
\description{
List the available layers (columns) in a CSV file.
}
