% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozflux.r
\name{available_layers_ozflux}
\alias{available_layers_ozflux}
\title{List all available layers for a particular output.}
\usage{
available_layers_ozflux(source, quant, sites = NULL)
}
\arguments{
\item{source:}{An OZFLUX source object.}

\item{quant:}{The quantity (ie LPJ-Guess output) to query.}

\item{sites:}{Optional list of site names to check. NULL means all sites.}
}
\value{
Returns a list of all layers (ie plottable columns) available for
this quantity (output).
}
\description{
List all available layers for a particular output.
}
\author{
Drew Holzworth
}
\keyword{internal}
