% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-csv.r
\name{available_quantities_csv}
\alias{available_quantities_csv}
\title{List all CSV *.csv files in a source directory}
\usage{
available_quantities_csv(source, names)
}
\arguments{
\item{source}{A \code{\linkS4class{Source}} containing the meta-data about
the CSV source.}

\item{names}{Logical, if TRUE return a character vector of names of
available quantities, if FALSE return a list of the actual Quantities.}
}
\value{
A list of all the .csv files present, with the ".csv" removed.#'
}
\description{
Simply lists all CSV output variables (stored as .csv files) available in a
directory.
}
