% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozflux.r
\name{available_quantities_ozflux}
\alias{available_quantities_ozflux}
\title{List all LPJ-GUESS *.out files which have been generated by all ozflux sites.}
\usage{
available_quantities_ozflux(source, names = TRUE, verbose = FALSE)
}
\arguments{
\item{source}{A GUESS source object}

\item{names}{Logical, if TRUE return the names of the quantities, if FALSE
return the quanties themseleves.}
}
\value{
A list of all the .out files present, with the ".out" removed.
}
\description{
The LPJ-GUESS dave_* branches have ozflux point-level simulations configured
in separate directories. This functions lists the outputs which have been
written by all site-level simulations.
}
\keyword{internal}
