% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-agb.r
\name{benchmark_agb}
\alias{benchmark_agb}
\title{Do Above-Ground Biomass benchmark.}
\usage{
benchmark_agb(params, settings, tables)
}
\arguments{
\item{params}{Benchmark parameters}

\item{settings}{Benchmark settings}

\item{tables}{Benchmark tables}
}
\value{
Returns a list containing the following named items:
\itemize{
\item maps: Max AGB in each gridcell.
\item trends: Temporal trend in AGB for each gridcell
\item comparisons: Output of fullSpatialComparison()
\item tables: Benchmark tables
\item benchmark: The benchmark object containing standard metadata
}
}
\description{
Do Above-Ground Biomass benchmark.
}
