% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-bom-lai.r, R/benchmark-auseflux.r
\name{benchmark_bom_lai}
\alias{benchmark_bom_lai}
\alias{benchmark_auseflux}
\title{Do BoM LAI benchmarks.}
\usage{
benchmark_bom_lai(settings, params, tables_list)

benchmark_auseflux(var, settings, params, tables_list)
}
\arguments{
\item{var}{Variable to be benchmarked. Should be one of: GPP, NEE, ER}

\item{settings:}{Benchmark settings.}

\item{params:}{Benchmark parameters.}

\item{tables_list:}{Benchmark tables.}
}
\value{
A DGVMTools::Field object containing the BoM LAI data.

A DGVMTools::Field object containing the BoM LAI data.
}
\description{
Do BoM LAI benchmarks.

Do AusEFlux benchmarks for a particular variable.
}
\author{
Drew Holzworth \email{d.holzworth@westernsydney.edu.au}
}
