% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-grasses.r
\name{benchmark_grass_dist}
\alias{benchmark_grass_dist}
\title{Do c3/c4 grass distribution benchmarks.}
\usage{
benchmark_grass_dist(settings, params, tables)
}
\arguments{
\item{settings:}{Benchmark settings.}

\item{params:}{Benchmark parameters.}

\item{tables:}{Benchmark tables.}
}
\value{
A list.
}
\description{
Do c3/c4 grass distribution benchmarks.
}
