% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-grasses.r
\name{benchmark_grasses}
\alias{benchmark_grasses}
\title{Do C3/C4 grass distribution benchmarks.}
\usage{
benchmark_grasses(settings, params, tables)
}
\arguments{
\item{settings}{Benchmark settings}

\item{params}{Benchmark parameters}

\item{tables}{Benchmark tables}
}
\value{
Returns a list containing the benchmark results, or NULL if DGVMBenchmarks is not available
}
\description{
Do C3/C4 grass distribution benchmarks.
}
