% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-lai.r
\name{benchmark_lai}
\alias{benchmark_lai}
\title{Do LAI benchmarks.}
\usage{
benchmark_lai(settings, params, tables, lai_type = "tot")
}
\arguments{
\item{settings}{Benchmark settings}

\item{params}{Benchmark parameters}

\item{tables}{Benchmark tables. This will be rbind-ed and returned as
result$tables.}

\item{lai_type}{"tot" for total LAI, "per" for persistent (ie trees), and
"rec" for recurrent (ie grasses).}
}
\value{
Returns a list containing the following named items:
\itemize{
\item maps: Mean maximum annual LAI for each dataset.
\item trends: Trend in LAI for each dataset.
\item seasonals: Mean LAI for each month of the year, for each dataset.
\item areas: Mean maximum annual LAI for each dataset.
\item comparisons: Spatial comparisons between datasets.
\item tables: Benchmark tables.
\item benchmark: Benchmark object.
}
}
\description{
This benchmark considers subannual LAI and will do seasonal analysis.
}
