% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation-reader.r
\name{clear_observation_registry}
\alias{clear_observation_registry}
\title{Clear the observation reader registry}
\usage{
clear_observation_registry()
}
\description{
This function is mainly intended for testing purposes.
}
