% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.r
\name{closest}
\alias{closest}
\title{For a given coordinate, return the closest coordinate from the list coords.}
\usage{
closest(coord, coords)
}
\arguments{
\item{coord}{A coordinate value.}

\item{coords}{List of coordinate values.}
}
\description{
For a given coordinate, return the closest coordinate from the list coords.
}
\keyword{internal}
