% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-lai-dist.r
\name{compare}
\alias{compare}
\title{Compare a layer between the specified fields.}
\usage{
compare(fields, i, j, lyr)
}
\arguments{
\item{fields}{List of fields.}

\item{i}{index of the first field to be compared (1-indexed)}

\item{j}{index of the second field to be compared (1-indexed)}
}
\description{
Compare a layer between the specified fields.
}
\keyword{internal}
