% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dave-stats.r
\name{compute_bias}
\alias{compute_bias}
\title{Compute bias of predictions to observations.}
\usage{
compute_bias(observations, predictions)
}
\arguments{
\item{observations}{Observed data.}

\item{predictions}{Predicted data.}
}
\description{
Compute bias of predictions to observations.
}
\keyword{internal}
