% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dave-stats.r
\name{compute_nse}
\alias{compute_nse}
\title{Compute nash-sutcliffe efficiency (-Inf - 1)}
\usage{
compute_nse(x, y)
}
\arguments{
\item{x:}{Observations}

\item{y:}{Predictions}
}
\description{
This is one minus the ratio of the error variance in the predictions to the
variance in the observations.
}
\details{
An efficiency of 1 means the predictions perfectly match the observations.
An efficiency of 0 means that the predictions are equally accurate predictors
as the mean of the observed data. An efficiency of < 0 means that the mean of
the observations is a better predictor than the model.
}
\author{
Drew Holzworth
}
\keyword{internal}
