% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dave-stats.r
\name{compute_rmse}
\alias{compute_rmse}
\title{Compute root mean square error.}
\usage{
compute_rmse(x, y)
}
\arguments{
\item{x}{Observations}

\item{y}{Predictions}
}
\description{
This is in the units of the variable.
}
\keyword{internal}
