% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dave-stats.r
\name{compute_rsr}
\alias{compute_rsr}
\title{Computes RSR of predictions to observed values.}
\usage{
compute_rsr(x, y)
}
\arguments{
\item{x}{Observations}

\item{y}{Predictions}
}
\description{
Compute ratio of RMSE to standard deviation of observations (0 - Inf).
}
\details{
Value of 0 means 0 RMSE (ie perfect fit).
}
\keyword{internal}
