% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.r
\name{convert_to_1degree}
\alias{convert_to_1degree}
\title{Convert the coordinate to the closest coordinate on a 1-degree grid.}
\usage{
convert_to_1degree(coord)
}
\arguments{
\item{coord:}{The coordinate value.}
}
\description{
Convert the coordinate to the closest coordinate on a 1-degree grid.
}
\keyword{internal}
