% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-csv.r
\name{convert_to_date}
\alias{convert_to_date}
\title{Convert a character vector of date strings to Date objects.}
\usage{
convert_to_date(date_strings, date_fmt)
}
\arguments{
\item{date_strings}{A character vector of date strings.}

\item{date_fmt}{The format of the date strings. E.g. "\%Y-\%m-\%d".}
}
\value{
A Date vector.
}
\description{
Convert a character vector of date strings to Date objects.
}
\keyword{internal}
