% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation-reader.r
\name{create_csv_reader}
\alias{create_csv_reader}
\title{Create a CSV file reader.}
\usage{
create_csv_reader(
  id,
  name,
  file_path,
  quant,
  layers,
  lat_col = "Lat",
  lon_col = "Lon",
  site_col = NULL,
  time_col = "date",
  infer_site = FALSE,
  time_fmt
)
}
\arguments{
\item{id}{Character. ID of the reader.}

\item{name}{Character. Name of the reader.}

\item{file_path}{Character. Path to the CSV file.}

\item{quant}{Character vector. The quantity encapsulated by this file.}

\item{layers}{Character vector. List of layers to read. If this is a named
vector, the names are used as the variable IDs.}

\item{lat_col}{Character. Name of the column containing latitude data.}

\item{lon_col}{Character. Name of the column containing longitude data.}

\item{site_col}{Character. Name of the column containing site names.}

\item{time_col}{Character. Name of the column containing time data.}

\item{infer_site}{Logical. Whether to infer site names from latitude and longitude or site_col.}

\item{time_fmt}{Character. The format of the time column. E.g. \%Y-\%m-\%d.}
}
\value{
An ObservationReader object.
}
\description{
Create a CSV file reader.
}
