% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation-reader.r
\name{create_flux_data_reader}
\alias{create_flux_data_reader}
\title{Create a NetCDF reader for the default ozflux-obs.nc file}
\usage{
create_flux_data_reader()
}
\value{
An ObservationReader object for the default NetCDF file.
}
\description{
Create a NetCDF reader for the default ozflux-obs.nc file
}
\keyword{internal}
