% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dave-config.r
\name{dave_config}
\alias{dave_config}
\title{Configure ozflux versions.}
\usage{
dave_config(versions, log_level, warning_as_error, log_file)
}
\arguments{
\item{versions:}{A list of DGVMTools::version objects.}

\item{log_level:}{Optional log level (0-4, default 2)
(\seealso{\link{set_log_level}})}

\item{log_file:}{Log file to be used. Empty string or NULL means stdout.}

\item{warning_as_error:}{Treat warnings as errors (FALSE by default)}
}
\description{
Configure ozflux versions.
}
\author{
Drew Holzworth \email{d.holzworth@westernsydney.edu.au}
}
