% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\docType{class}
\name{DaveBenchmark-class}
\alias{DaveBenchmark-class}
\title{Dave Benchmark class}
\description{
This is the class used for all DAVE benchmarks. It extends the base Benchmark
class provided by the DGVMBenchmarks package, because the class exported from
that package is missing a few crucial fields.
}
\section{Slots}{

\describe{
\item{\code{simulation_format}}{Format of the simulation}

\item{\code{dataset_source}}{Reference/source of the dataset}
}}

