% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tellaus.r
\name{do_tellaus_plots}
\alias{do_tellaus_plots}
\title{Produce a standard set of plots for the tellaus report from the given
benchmark.}
\usage{
do_tellaus_plots(benchmark, settings, params, overall_tables, title)
}
\arguments{
\item{benchmark}{A list with the following elements:
- maps: Temporally-aggregated dataset (one value per grid cell)
- trends: Temporal trends (one value per grid cell)
- seasonals: Seasonally-aggregated values (one value per month per gridcell)
- comparisons: Output of DGVMTools::fullSpatialComparison()
- tables: List with two named data frames: "totals" and "metrics"
- benchmark: A DaveBenchmark instance}

\item{settings}{Benchmarking settings as specified by the user}

\item{params}{Benchmarking parameters as specified by the user}

\item{overall_tables}{List of all tables}
}
\value{
A list of tags
}
\description{
Produce a standard set of plots for the tellaus report from the given
benchmark.
}
\author{
Drew Holzworth \email{d.holzworth@westernsydney.edu.au}
}
