% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-csv.r
\name{filter_sites}
\alias{filter_sites}
\title{Filter a data.table by site}
\usage{
filter_sites(dt, sites, lat_col, lon_col, site_col)
}
\arguments{
\item{dt}{A data.table to filter}

\item{sites}{A character vector of site names to keep}

\item{lat_col}{The name of the column containing latitude data}

\item{lon_col}{The name of the column containing longitude data}

\item{site_col}{The name of the column containing site names}
}
\value{
A data.table filtered by site
}
\description{
Filter a data.table by site
}
\keyword{internal}
