% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-csv.r
\name{filter_stainfo}
\alias{filter_stainfo}
\title{Get the list of sites defined by a spatial extent.}
\usage{
filter_stainfo(target_stainfo)
}
\arguments{
\item{target_stainfo}{An STAInfo object defining the spatial-temporal-annual extent over which we want the data}
}
\value{
A character vector of site names
}
\description{
Get the list of sites defined by a spatial extent.
}
\keyword{internal}
