% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation-reader.r
\name{find_readers_for_var}
\alias{find_readers_for_var}
\title{Find readers that support a given variable}
\usage{
find_readers_for_var(var_id)
}
\arguments{
\item{var_id}{Character. ID of the variable to find readers for}
}
\value{
A list of readers that support the variable
}
\description{
Find readers that support a given variable
}
