% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tellaus.r
\name{fix_xaxis_label_overlap}
\alias{fix_xaxis_label_overlap}
\title{Fix overlapping longitude values on the x-axis of a spatial plot.}
\usage{
fix_xaxis_label_overlap(plt)
}
\arguments{
\item{plt}{A plot}
}
\value{
\link{ggplot} A plot with fixed x-axis labels.
}
\description{
Fix overlapping longitude values on the x-axis of a spatial plot.
}
\keyword{internal}
