% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-lai.r
\name{get_bom_data_path}
\alias{get_bom_data_path}
\title{Get the location of the BoM LAI data.}
\usage{
get_bom_data_path(data_path, type)
}
\arguments{
\item{data_path}{Path to data.}

\item{type}{Should be either "rec", "per", or "tot"}

\item{author}{Drew Holzworth}
}
\value{
Path to the NetCDF file containing BoM LAI data.
}
\description{
Get the location of the BoM LAI data.
}
\keyword{internal}
