% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-lai.r
\name{get_bom_location}
\alias{get_bom_location}
\title{Get the location of the BoM LAI data.}
\usage{
get_bom_location(data_path, filename)
}
\arguments{
\item{data_path}{Path to the data directory provided alongside this package.}

\item{filename}{Filename, which depends on whether we want understory,
overstory, or total LAI.}
}
\value{
Path to the NetCDF file containing BoM LAI data
}
\description{
Get the location of the BoM LAI data.
}
\keyword{internal}
