% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-bom-lai.r
\name{get_bom_understory_location}
\alias{get_bom_understory_location}
\title{Get the location of the BoM understory LAI data that is shipped with this
package.}
\usage{
get_bom_understory_location(data_path)
}
\arguments{
\item{author}{Drew Holzworth}
}
\value{
Path to the NetCDF file containing BoM LAI data.
}
\description{
Get the location of the BoM understory LAI data that is shipped with this
package.
}
\keyword{internal}
