% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tellaus.r
\name{get_colour}
\alias{get_colour}
\title{Get a colour representing the goodness of a metric value.}
\usage{
get_colour(metric, value)
}
\arguments{
\item{metric}{Metric name (e.g. NSE, RMSE, etc).}

\item{value}{Metric value.}
}
\value{
\link{character} A colour representing the goodness of a particular value.
}
\description{
Get a colour representing the goodness of a metric value.
}
\keyword{internal}
