% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-csv.r
\name{get_file_path}
\alias{get_file_path}
\title{Get the file name for a given quantity}
\usage{
get_file_path(source, quant)
}
\arguments{
\item{source}{A \code{\linkS4class{Source}} containing the meta-data about
the CSV source.}

\item{quant}{A \code{\linkS4class{Quantity}} object to specify what quantity
should be opened.}
}
\value{
A character string holding the name of the file.
}
\description{
Get the file name for a given quantity
}
\keyword{internal}
