% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-agb.r
\name{get_fullcam_filepath}
\alias{get_fullcam_filepath}
\title{Get the path to the fullcam data file from the specified base data path.}
\usage{
get_fullcam_filepath(data_directory, highres = FALSE)
}
\arguments{
\item{highres}{True to read the high-resolution (0.0025°) dataset. False to
read the low-resolution (1°) dataset.}

\item{data_path}{Base path to data files.}
}
\value{
Returns a string containing the expected path to the fullcam AGB data
file, relative to the base data directory path.
}
\description{
Get the path to the fullcam data file from the specified base data path.
}
\keyword{internal}
