% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dave-globals.r
\name{get_global}
\alias{get_global}
\title{Get the package-internal global variable with the specified name.}
\usage{
get_global(name)
}
\arguments{
\item{name:}{Name of the variable.}
}
\value{
Returns the value of the variable
}
\description{
Get the package-internal global variable with the specified name.
}
\keyword{internal}
