% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dave-processing.r
\name{get_gridcell}
\alias{get_gridcell}
\title{Extract data for the specified gridcell.}
\usage{
get_gridcell(data, lat, lon, site_name = NULL)
}
\arguments{
\item{data:}{The input data as obtained from \seealso{\link{read_data}}}

\item{lat:}{Latitude of the gridcell}

\item{lon:}{Longitude of the gridcell.}

\item{site_name:}{Name of the ozflux site (optional).}
}
\value{
Returns a \seealso{\link{DGVMTools::Source}} object containing data
only for the specified location.
}
\description{
Extract data for the specified gridcell.
}
\author{
Drew Holzworth
}
\keyword{internal}
