% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.r
\name{get_grouping_names}
\alias{get_grouping_names}
\title{Get names of columns by which the data frame should be grouped when
aggregating spatially.}
\usage{
get_grouping_names(df)
}
\arguments{
\item{df}{A data frame.}
}
\description{
Get names of columns by which the data frame should be grouped when
aggregating spatially.
}
\keyword{internal}
