% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-sanitisation.r
\name{get_known_quantity}
\alias{get_known_quantity}
\title{Find a known quantity whose lpj-guess "standard" output file name is given by
name.}
\usage{
get_known_quantity(name)
}
\arguments{
\item{name}{Standard name of the output file.}
}
\value{
A \link[DGVMTools:Quantity]{DGVMTools::Quantity} object, or NULL if not found.
}
\description{
Find a known quantity whose lpj-guess "standard" output file name is given by
name.
}
\keyword{internal}
