% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dave-logging.r
\name{get_log_level}
\alias{get_log_level}
\title{Get the current log level used by dave_analysis logging functions.}
\usage{
get_log_level()
}
\value{
Log level (0-4)
}
\description{
Get the current log level used by dave_analysis logging functions.
}
\keyword{internal}
