% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-lai.r
\name{get_modis_data_path}
\alias{get_modis_data_path}
\title{Get the location of the MODIS LAI data.}
\usage{
get_modis_data_path(data_path)
}
\arguments{
\item{data_path}{Path to the data directory.}
}
\value{
Path to the MODIS LAI data file.
}
\description{
Get the location of the MODIS LAI data.
}
\keyword{internal}
