% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dave-processing.r
\name{get_observed_vars}
\alias{get_observed_vars}
\title{Get the list of temporal plottable variables from the observed .nc file.}
\usage{
get_observed_vars()
}
\value{
Returns a list of \seealso{\link{DGVMTools::Quantity}} objects.
}
\description{
Get the list of temporal plottable variables from the observed .nc file.
}
\author{
Drew Holzworth
}
\keyword{internal}
