% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation-reader.r
\name{get_reader}
\alias{get_reader}
\title{Get a reader by ID}
\usage{
get_reader(id)
}
\arguments{
\item{id}{Character. ID of the reader to get}
}
\value{
The reader object or NULL if not found
}
\description{
Get a reader by ID
}
