% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation-reader.r
\name{get_reader_names}
\alias{get_reader_names}
\title{Get the names of all registered observation readers.}
\usage{
get_reader_names()
}
\value{
A character vector of reader IDs
}
\description{
Get the names of all registered observation readers.
}
