% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-csv.r
\name{get_site_names_by_coord}
\alias{get_site_names_by_coord}
\title{Get site names from their latitudes and longitudes.}
\usage{
get_site_names_by_coord(lats, lons, all_sites, max_distance = 0.1)
}
\arguments{
\item{lats}{The latitude of the sites}

\item{lons}{The longitude of the sites}

\item{all_sites}{A data.table containing the latitude, longitude and name of all sites}

\item{max_distance}{Maximum allowable distance between coordinates (default: 0.1 degrees)
If the closest site is farther than this distance, returns NULL}
}
\value{
Character string with site name or NULL if no site is within max_distance
}
\description{
Get site names from their latitudes and longitudes.
}
\keyword{internal}
