% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dave-processing.r
\name{has_output}
\alias{has_output}
\title{Check if the given layer exists in the given source.}
\usage{
has_output(source, layer)
}
\arguments{
\item{source}{A DGVMTools::Source object}

\item{layer}{Name of a layer (e.g. "mlai")}
}
\description{
Check if the given layer exists in the given source.
}
\keyword{internal}
