% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozflux-benchmarks.r
\name{is_installed}
\alias{is_installed}
\title{Check if a package is installed.}
\usage{
is_installed(pkg)
}
\arguments{
\item{pkg}{Name of the package.}
}
\value{
Returns TRUE if the package is installed, otherwise FALSE.
}
\description{
Check if a package is installed.
}
\keyword{internal}
