% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-sanitisation.r
\name{is_known_quantity}
\alias{is_known_quantity}
\title{Check if a quantity is a known quantity.}
\usage{
is_known_quantity(name)
}
\arguments{
\item{name}{Name of a quantity.}
}
\value{
TRUE iff the quantity is a known quantity.
}
\description{
Check if a quantity is a known quantity.
}
\keyword{internal}
