% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation-reader.r
\name{list_all_readers}
\alias{list_all_readers}
\title{Get all registered observation readers.}
\usage{
list_all_readers()
}
\value{
List of ObservationReader objects.
}
\description{
Get all registered observation readers.
}
