% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation-reader.r
\name{list_readers}
\alias{list_readers}
\title{Get the names of all registered observation readers.}
\usage{
list_readers()
}
\value{
A character vector of reader IDs
}
\description{
Get the names of all registered observation readers.
}
