% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation-reader.r
\docType{class}
\name{ObservationReader-class}
\alias{ObservationReader-class}
\title{ObservationReader class}
\description{
A class representing a reader for observation data. This class provides a
standardized interface for reading observation data from various sources.
}
\section{Slots}{

\describe{
\item{\code{src}}{Source. A DGVMTools::Source object.}

\item{\code{file_path}}{Character. Path to the data file.}

\item{\code{available_quantities}}{Function. Get available quantities.}

\item{\code{available_layers}}{Function. Get available layers.}

\item{\code{read_func}}{Function. Function to read the data.}
}}

