% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozflux-benchmarks.r
\name{ozflux_benchmarks}
\alias{ozflux_benchmarks}
\title{Generate ozflux benchmarks.}
\usage{
ozflux_benchmarks(
  sources,
  use_plotly = FALSE,
  text_multiplier = 1.5,
  combined_graph = TRUE,
  max_title = 4,
  marker_size = 1
)
}
\arguments{
\item{use_plotly:}{Iff true, plots will be rendered using plotly. Note that
plotly plots are only interactive when rendering to HTML; therefore, we only
turn on plotly when rendering to HTML.This can of course be overriden (e.g.
for debugging).}

\item{combined_graph:}{Iff true, plots will be rendered in 2x2 panels for
each variable for each site.}

\item{max_title:}{Maximum number of nested titles.}

\item{marker_size:}{Marker size. Decrease for smaller markers.}
}
\value{
Returns a list of HTML tags if use_plotly is TRUE. Otherwise all
content will be printed, for use in R markdown.
}
\description{
Generate ozflux benchmarks.
}
\author{
Drew Holzworth
}
