% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozflux-plotting.r
\name{ozflux_panel}
\alias{ozflux_panel}
\title{Create a 2x2 panel of plots, identical to what is seen in the ozflux
benchmarks.}
\usage{
ozflux_panel(
  sources,
  var,
  sites = NULL,
  use_plotly = FALSE,
  common_yaxis = FALSE,
  show_all_observations = TRUE,
  show_all_predictions = TRUE
)
}
\arguments{
\item{sources}{Source or list of data sources to be plotted (see: \link{sanitise_source})}

\item{var}{The variable to be plotted (see: \link{sanitise_variable})}

\item{sites}{Ozflux sites to be plotted (see: \link{sanitise_ozflux_sites})}

\item{use_plotly}{iff TRUE, generate \link{plotly} outputs, FALSE to use \link[ggplot2:ggplot]{ggplot2::ggplot}}

\item{common_yaxis}{iff TRUE, use the same y-axis scales for all sites. FALSE
otherwise}

\item{show_all_observations:}{If true, all observations will be returned (ie the predicted data may contain NA values). If false, only the observations which have a matching prediction will be returned.}

\item{show_all_predictions:}{If true, all predictions will be returned (ie the observed data may contain NA values). If false, only the predictions which have a matching observation will be returned.}
}
\value{
Returns a single \link[ggplot2:ggplot]{ggplot2::ggplot} object if plotting one site, otherwise returns
a list of \link[ggplot2:ggplot]{ggplot2::ggplot} objects of the same length as the number of sites to be
plotted. If use_plotly is TRUE, returns \link{plotly} objects instead.
}
\description{
One series will be plotted on each graph for each source. One
panel will be created for each site.
}
