% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozflux-plotting.r
\name{ozflux_plot}
\alias{ozflux_plot}
\title{Create timeseries plots of predictions against observations.}
\usage{
ozflux_plot(
  sources,
  vars,
  sites = NULL,
  separate = FALSE,
  use_plotly = FALSE,
  common_yaxis = FALSE
)
}
\arguments{
\item{sources:}{One or more sources. These can be paths to lpj-guess
repositories, or \seealso{\link{DGVMTools::Source}} objects. This argument is
optional if sources have been configured via \seealso{\link{dave_config}}.}

\item{vars:}{The variables to be plotted. These may be specified as strings
containing the output file names without extension (e.g. "dave_lai"), or as
\seealso{\link{DGVMTools::Quantity}} objects.}

\item{sites:}{List of sites to be plotted. These may be specified as either
strings (site names), or as tuples of (lon, lat). If NULL is provided, all
ozflux sites will be plotted.}

\item{separate:}{Iff TRUE, a list of the site-level plots will be returned.
If FALSE, all plots will be rendered to a single panel. This has no effect
if a single site is to be plotted.}

\item{use_plotly:}{Iff true, graphs will be plotted using plotly.}

\item{common_yaxis:}{Iff true, all plots in the panel will have the same y-
axis range. No effect if only 1 site is being plotted.}
}
\value{
Can return a single, or vector of, ggplot or plotly objects.
}
\description{
Plots a single variable across predictions at each specified ozflux site,
against observations (if any are available), generating one plot per site. If
separate is TRUE, then a separate plot will be returned for each site. If
separate is FALSE, then a panel of all plots will be returned.
}
\details{
If data is missing for any of the sites to be plotted, an error will occur.
}
\author{
Drew Holzworth
}
