% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozflux-plotting.r
\name{ozflux_plot_layerwise}
\alias{ozflux_plot_layerwise}
\title{Create ozflux site-level layerwise plots.}
\usage{
ozflux_plot_layerwise(
  sources,
  var,
  sites,
  layers = NULL,
  title = NULL,
  separate_plots = TRUE,
  use_plotly = FALSE
)
}
\arguments{
\item{sources:}{Sources of data. Should be paths to lpj-guess repos or
\seealso{\link{DGVMTools::Source}} objects.}

\item{var:}{The variable to be plotted, specified as the output filename
without extension (e.g. `"dave_sw"`), or as
\seealso{\link{DGVMTools::Quantity}} objects.}

\item{sites:}{Ozflux sites to be plotted, expressed as site names or as
tuples of (lon, lat).}

\item{layers:}{The layers to be plotted. E.g. `paste0("sw_", 0:14)`.}

\item{title:}{The desired panel titles.}

\item{separate_plots:}{TRUE to draw each layer in a separate plot. FALSE to
draw all layers on the same plot.}

\item{use_plotly:}{TRUE to draw plots with plotly. FALSE to use ggplot2.}
}
\value{
Returns a list of ggplot objects.
}
\description{
Create a panel of plots, where each plot in the panel contains one of the
layers of the specified variable. One panel will be generated for each site.
}
\details{
(In this nomenclature, a "layer" technically refers to a column in the output
file).
}
\author{
Drew Holzworth
}
