% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tellaus.r
\name{plot_absolute_deltas}
\alias{plot_absolute_deltas}
\title{Create an absolute values differenes plot panel for use in the tellaus
benchmarks.}
\usage{
plot_absolute_deltas(comparisons, settings)
}
\value{
A list of tags
}
\description{
Create an absolute values differenes plot panel for use in the tellaus
benchmarks.
}
\author{
Drew Holzworth \email{d.holzworth@westernsydney.edu.au}
}
