% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tellaus.r
\name{plot_absolute_values}
\alias{plot_absolute_values}
\title{Create a spatial plot of the absolute values of the given dataset.}
\usage{
plot_absolute_values(benchmark, settings, limits = NULL)
}
\arguments{
\item{benchmark}{The benchmark, the output of one of the benchmark_*
functions included in this package. Should be a list.}

\item{settings}{A list of benchmarking settings as in tellaus.rmd.}
}
\value{
A ggplot object.
}
\description{
Create a spatial plot of the absolute values of the given dataset.
}
\keyword{internal}
