% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozflux-plotting.r
\name{plot_layerwise}
\alias{plot_layerwise}
\title{Create layerwise plots.}
\usage{
plot_layerwise(sites, sources, var, layers, title = NULL)
}
\arguments{
\item{sites:}{Ozflux sites to be plotted, expressed as tuples of (lon, lat).}

\item{sources:}{Sources of data. Should be paths to lpj-guess repos.}

\item{var:}{The variable to be plotted. E.g. `"dave_sw"``.}

\item{layers:}{The layers to be plotted. E.g. `paste0("sw_", 0:14)`.}
}
\value{
Returns a list of ggplot objects.
}
\description{
Create a panel of plots, where each panel contains one of the layers of the
specified variable. One panel will be generated for each site.
}
\author{
Drew Holzworth
}
