% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozflux-plotting-internal.r
\name{plot_temporal}
\alias{plot_temporal}
\title{Plot temporal data}
\usage{
plot_temporal(
  fields,
  layers = NULL,
  gridcells = NULL,
  title = character(0),
  subtitle = character(0),
  cols = NULL,
  col.by = "Layer",
  col.labels = waiver(),
  linetypes = NULL,
  linetype.by = NULL,
  linetype.labels = waiver(),
  linewidths = NULL,
  linewidth.by = NULL,
  linewidth.labels = waiver(),
  sizes = NULL,
  size.by = NULL,
  size.labels = waiver(),
  shapes = NULL,
  shape.by = NULL,
  shape.labels = waiver(),
  alphas = NULL,
  alpha.by = NULL,
  alpha.labels = waiver(),
  y.label = NULL,
  y.lim = NULL,
  x.label = NULL,
  x.lim = NULL,
  legend.position = "bottom",
  text.multiplier = NULL,
  dropEmpty = FALSE,
  plotTrend = FALSE,
  ...
)
}
\value{
A ggplot
}
\description{
This is a modified copy of \link[DGVMTools:plotTemporal,]{DGVMTools::plotTemporal,} with the ability to plot
some fields as points instead of lines.
}
\details{
For more information, see: \link[DGVMTools:plotTemporal]{DGVMTools::plotTemporal}
}
\keyword{internal}
