% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation-reader.r
\name{populate_registry}
\alias{populate_registry}
\title{Populate the observation reader registry with a default set of readers.}
\usage{
populate_registry()
}
\description{
Populate the observation reader registry with a default set of readers.
}
\keyword{internal}
