% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-auseflux.r
\name{read_auseflux}
\alias{read_auseflux}
\title{Read AusEFlux data from the specified data file, or from the default location if
no data file is provided.}
\usage{
read_auseflux(var, data_path)
}
\arguments{
\item{var}{Variable to read (see: \link{auseflux_sanitise_var})}

\item{data_path}{Path to the data directory}
}
\value{
Returns a \link[DGVMTools:Field]{DGVMTools::Field} object containing the requested AusEFlux data
}
\description{
Read AusEFlux data from the specified data file, or from the default location if
no data file is provided.
}
