% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-lai-dist.r
\name{read_bom_tree_dist}
\alias{read_bom_tree_dist}
\title{Read BoM LAI estimates.}
\usage{
read_bom_tree_dist(data_path)
}
\arguments{
\item{data_path}{Path to data directory.}
}
\description{
Read BoM LAI estimates and post-process them to calculate species
distribution.
}
\author{
Drew Holzworth \email{d.holzworth@westernsydney.edu.au}
}
\keyword{internal}
