% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dave-processing.r
\name{read_data}
\alias{read_data}
\title{Read data for the specified variable.}
\usage{
read_data(vars, sources, site = NULL, layers = NULL)
}
\arguments{
\item{vars:}{The variables (a list of DGVMTools::Quantity objects).}

\item{sources:}{Input sources (\seealso{\link{sanitise_sources}}}

\item{site:}{Optional name of the ozflux site for which data should be read.
If NULL, data for all sites will be returned.}

\item{layers:}{Names of layers to be read.}
}
\value{
A single DGVMTools::Field with a layer of observations, and one layer
for each source specified in sources.
}
\description{
Read data for the specified variable.
}
\author{
Drew Holzworth
}
\keyword{internal}
