% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-agb.r
\name{read_fullcam}
\alias{read_fullcam}
\title{Read Fullcam potential AGB data from the specified data path.}
\usage{
read_fullcam(data_directory, highres = FALSE)
}
\arguments{
\item{highres}{True to read the high-resolution (0.0025°) dataset. False to
read the low-resolution (1°) dataset.}

\item{data_path}{Base path to data files.}
}
\value{
Returns a \link[DGVMTools:Field]{DGVMTools::Field} object containing fullcam AGB data.
}
\description{
Read Fullcam potential AGB data from the specified data path.
}
\keyword{internal}
