% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-gpp.r
\name{read_gosif}
\alias{read_gosif}
\title{Read GOSIF GPP from the specified data path.}
\usage{
read_gosif(data_path)
}
\arguments{
\item{data_path}{Path to data files.}
}
\value{
A \link[DGVMTools:Field]{DGVMTools::Field} object containing GOSIF global monthly GPP.
}
\description{
Read GOSIF GPP from the specified data path.
}
\keyword{internal}
