% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-lai-dist.r
\name{read_guess}
\alias{read_guess}
\title{Read lpj-guess LAI outputs.}
\usage{
read_guess(src, fields)
}
\arguments{
\item{src}{A \link[DGVMTools:Source]{DGVMTools::Source} object.}
}
\description{
This function reads LPJ-GUESS outputs and post-processes them to calculate
species distribution.
}
\keyword{internal}
