% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-lai.r
\name{read_modis_lai}
\alias{read_modis_lai}
\title{Read MODIS LAI from the specified data path.}
\usage{
read_modis_lai(data_path)
}
\arguments{
\item{data_path}{Path to the data directory}
}
\value{
Returns a \link[DGVMTools:Field]{DGVMTools::Field} object containing the MODIS LAI data
}
\description{
Read MODIS LAI from the specified data path.
}
