% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation-reader.r
\name{read_observation}
\alias{read_observation}
\alias{read_observation,ObservationReader,character-method}
\title{Read observation data}
\usage{
read_observation(reader, var_id, ...)

\S4method{read_observation}{ObservationReader,character}(reader, var_id, ...)
}
\arguments{
\item{reader}{An ObservationReader object.}

\item{var_id}{Character. ID of the variable to read.}

\item{...}{Additional arguments passed to the reader function.}
}
\value{
A DGVMTools::Field object containing the observation data.
}
\description{
Read observation data for a specific variable using the provided reader.
}
