% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-sanitisation.r
\name{read_ozflux_sites}
\alias{read_ozflux_sites}
\title{Get a list of all known ozflux sites.}
\usage{
read_ozflux_sites()
}
\value{
Data table containing three columns: Lon, lat, Name.
}
\description{
Get a list of all known ozflux sites.
}
\author{
Drew Holzworth
}
