% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation-reader.r
\name{register_csv_reader}
\alias{register_csv_reader}
\title{Create a CSV file reader.}
\usage{
register_csv_reader(
  id,
  name,
  file_path,
  quant,
  layers,
  lat_col = "Lat",
  lon_col = "Lon",
  site_col = NULL,
  time_col = "date"
)
}
\arguments{
\item{file_path}{Character. Path to the CSV file.}

\item{quant}{Character vector. The quantity encapsulated by this file.}

\item{layers}{Character vector. List of layers to read. If this is a named
vector, the names are used as the variable IDs.}
}
\value{
An ObservationReader object.
}
\description{
Create a CSV file reader.
}
