% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation-reader.r
\name{register_supplementary_observations}
\alias{register_supplementary_observations}
\title{Register supplementary observation readers for the "standard" set of
observations which are not included in this package.}
\usage{
register_supplementary_observations(obs_dir)
}
\arguments{
\item{obs_dir}{Character. Path to the observation directory.}
}
\description{
Register supplementary observation readers for the "standard" set of
observations which are not included in this package.
}
