% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-sanitisation.r
\name{sanitise_source}
\alias{sanitise_source}
\title{Sanitise version inputs for plotting. Returns a list of \link[DGVMTools:Source]{DGVMTools::Source}
objects.}
\usage{
sanitise_source(source)
}
\arguments{
\item{source:}{May be one of:
\itemize{
\item A [\link[DGVMTools:Source]{DGVMTools::Source}] object
\item Path to an lpj-guess repository
\item Path to a directory containing LPJ-Guess output files
}}
}
\value{
Returns a list of [\link[DGVMTools:Source]{DGVMTools::Source}] objects.
}
\description{
This is an internal helper function and is not intended for external use.
}
\keyword{internal}
