% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-sanitisation.r
\name{sanitise_sources}
\alias{sanitise_sources}
\title{Sanitise version inputs for plotting. Returns a list of DGVMTools::Source
objects.}
\usage{
sanitise_sources(sources)
}
\arguments{
\item{sources:}{May be one of:
    - A single DGVMTools::Source object
    - List of DGVMTools::Source objects
    - A single string providing path to an lpj-guess repository,
      or a subdirectory of an lpj-guess repository
- A list of paths to lpj-guess repositories}
}
\value{
Returns a list of Sources.
}
\description{
This is an internal helper function and is not intended for external use.
}
\author{
Drew Holzworth \email{d.holzworth@westernsydney.edu.au}
}
\keyword{internal}
