% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-sanitisation.r
\name{sanitise_spatial_extent}
\alias{sanitise_spatial_extent}
\title{Get the canonical name of an ozflux site.}
\usage{
sanitise_spatial_extent(spatial_extent, sites)
}
\arguments{
\item{spatial_extent:}{The spatial extent representing the site, supplied as
a raster::extent object or an object from which a
raster::extent object can be derived - eg. a Raster*
object or another Field object.}
}
\value{
This function is not yet implemented and will raise an error.
}
\description{
Return the name of the ozflux site specified by the given spatial extent.
Raises an error if the site name is invalid.
}
\keyword{internal}
