% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-sanitisation.r
\name{sanitise_spatial_extent_id}
\alias{sanitise_spatial_extent_id}
\title{Get the canonical name of an ozflux site.}
\usage{
sanitise_spatial_extent_id(spatial_extent_id, sites)
}
\arguments{
\item{spatial_extent_id:}{Name of the site specified by the user.}

\item{sites:}{List of valid sites.}
}
\value{
Returns the name of the ozflux site.
}
\description{
Return the name of the specified ozflux site, or raise an error if the site
name is invalid.
}
\keyword{internal}
