% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-sanitisation.r
\name{sanitise_variable}
\alias{sanitise_variable}
\title{Sanitise a variable provided by the user.}
\usage{
sanitise_variable(var)
}
\arguments{
\item{var:}{May be one of:
\itemize{
\item A [\link[DGVMTools:Quantity]{DGVMTools::Quantity}] object (for full control over units/name/filename)
\item File name without extension of an lpj-guess output (e.g. lai)
}}
}
\value{
Returns a [\link[DGVMTools:Quantity]{DGVMTools::Quantity}] object.
}
\description{
Sanitise a variable provided by the user.
}
\keyword{internal}
