% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dave-globals.r
\name{set_global}
\alias{set_global}
\title{Set the package-internal global variable with the specified name.}
\usage{
set_global(name, value)
}
\arguments{
\item{name}{Name of the variable.}

\item{value}{Value of the variable.}
}
\description{
Set the package-internal global variable with the specified name.
}
\keyword{internal}
