% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dave-logging.r
\name{set_log_level}
\alias{set_log_level}
\title{Set the log level used by dave_analysis logging functions.}
\usage{
set_log_level(level)
}
\description{
This should be in range 0-4. Higher values will result in more log messages
being written.
}
\author{
Drew Holzworth
}
