% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation-reader.r
\name{show,ObservationReader-method}
\alias{show,ObservationReader-method}
\title{Show method for ObservationReader}
\usage{
\S4method{show}{ObservationReader}(object)
}
\arguments{
\item{object}{An ObservationReader object.}
}
\value{
None. Prints information about the reader to the console.
}
\description{
Show method for ObservationReader
}
