% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-lai-dist.r
\name{sum_pfts}
\alias{sum_pfts}
\title{Calculate the sum of the specified columns.}
\usage{
sum_pfts(df, pfts)
}
\arguments{
\item{df}{A data frame.}

\item{pfts}{Names of columns to be summed.}
}
\description{
Calculate the sum of the specified columns.
}
\keyword{internal}
