% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation-reader.r
\name{supports_variable}
\alias{supports_variable}
\alias{supports_variable,ObservationReader,character-method}
\title{Check if a variable is supported by a reader}
\usage{
supports_variable(reader, var_id)

\S4method{supports_variable}{ObservationReader,character}(reader, var_id)
}
\arguments{
\item{reader}{An ObservationReader object.}

\item{var_id}{Character. ID of the variable to check.}
}
\value{
Logical. TRUE if the variable is supported, FALSE otherwise.
}
\description{
Check if a variable is supported by a reader
}
