% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-agb.r
\name{validate_agb}
\alias{validate_agb}
\title{Ensure that the required files for the Above-Ground Biomass benchmark exist
in the given data directory.}
\usage{
validate_agb(data_directory, highres = FALSE)
}
\arguments{
\item{data_directory}{Base path to data files.}

\item{highres}{True to read the high-resolution (0.0025°) dataset. False to
read the low-resolution (1°) dataset.}
}
\description{
Ensure that the required files for the Above-Ground Biomass benchmark exist
in the given data directory.
}
