% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-auseflux.r
\name{validate_auseflux}
\alias{validate_auseflux}
\title{Ensure that all data exists which is required for this benchmark.}
\usage{
validate_auseflux(data_path, var)
}
\arguments{
\item{data_path}{Path to the data directory}
}
\description{
Ensure that all data exists which is required for this benchmark.
}
