% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-lai.r
\name{validate_lai}
\alias{validate_lai}
\title{Ensure that all data exists which is required for this benchmark.}
\usage{
validate_lai(data_path)
}
\arguments{
\item{data_path}{Path to the data directory}
}
\description{
Ensure that all data exists which is required for this benchmark.
}
