% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-gpp.r
\name{benchmark_bom_lai}
\alias{benchmark_bom_lai}
\alias{benchmark_gpp}
\title{Do GPP benchmark, using Auseflux and GOSIF GPP products. The GOSIF GPP will
be cropped to only contain Australia.}
\usage{
benchmark_gpp(params, settings, tables)
}
\arguments{
\item{params}{Benchmark parameters}

\item{settings}{Benchmark settings}

\item{tables}{Benchmark tables}
}
\value{
Returns a list containing the following named items:
\itemize{
\item maps: Mean of yearly total GPP for each gridcell
\item trends: Temporal trend in GPP for each gridcell
\item seasonals: Gridded mean monthly GPP for each month over the time period
\item comparisons: Output of fullSpatialComparison()
\item tables: Benchmark tables
\item benchmark: The benchmark object containing standard metadata
}
}
\description{
Do GPP benchmark, using Auseflux and GOSIF GPP products. The GOSIF GPP will
be cropped to only contain Australia.
}
