% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-lai.r, R/benchmark-auseflux.r,
%   R/benchmark-gpp.r
\name{benchmark_bom_lai}
\alias{benchmark_bom_lai}
\alias{benchmark_lai}
\alias{benchmark_auseflux}
\alias{benchmark_gpp}
\title{Do LAI benchmarks.}
\usage{
benchmark_lai(settings, params, tables)

benchmark_auseflux(var, settings, params, tables)

benchmark_gpp(params, settings, tables)
}
\arguments{
\item{var}{Variable to be benchmarked. Should be one of: GPP, NEE, ER}

\item{params:}{Benchmark parameters.}

\item{settings:}{Benchmark settings.}

\item{tables:}{Benchmark tables.}
}
\value{
A \link[DGVMTools:Field]{DGVMTools::Field} object containing the BoM LAI data.

A \link[DGVMTools:Field]{DGVMTools::Field} object containing the BoM LAI data.

A list containing the following named items:
\itemize{
\item maps: Mean of yearly total GPP for each gridcell
\item trends: Temporal trend in GPP for each gridcell
\item seasonals: Gridded mean monthly GPP for each month over the time period
\item comparisons: Output of fullSpatialComparison()
\item tables: Benchmark tables
\item benchmark: The benchmark object containing standard metadata
}
}
\description{
This benchmark considers subannual LAI and will do seasonal analysis.
}
