% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-lai-dist.r
\name{compare_all}
\alias{compare_all}
\title{Compare a layer between all combinations of the fields.}
\usage{
compare_all(fields, lyr)
}
\arguments{
\item{fields}{The fields to be compared.}

\item{lyr}{The layer to be compared in the fields. Must exist in all fields.}
}
\description{
E.g. for 3 fields: compare fields 1&2, 1&3, 2&3.
}
\author{
Drew Holzworth \email{d.holzworth@westernsydney.edu.au}
}
