% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dave-config.r
\name{dave_config}
\alias{dave_config}
\title{Configure ozflux versions.}
\usage{
dave_config(versions, log_level, warning_as_error, log_file)
}
\arguments{
\item{versions}{A list of \link[DGVMTools:version]{DGVMTools::version} objects}

\item{log_level}{Optional log level (0-4, default 2) (see: \link{set_log_level})}

\item{warning_as_error}{iff TRUE, treat warnings as errors (default: FALSE)}

\item{log_file}{Log file to be used. Empty string or NULL means stdout}
}
\description{
Configure ozflux versions.
}
