% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tellaus.r
\name{do_tellaus_plots}
\alias{do_tellaus_plots}
\title{Produce a standard set of plots for the tellaus report from the given
benchmark.}
\usage{
do_tellaus_plots(benchmark, settings, params, overall_tables, title)
}
\arguments{
\item{benchmark}{A list with the following elements:
\itemize{
\item maps: Temporally-aggregated dataset (one value per grid cell)
\item trends: Temporal trends (one value per grid cell)
\item seasonals: Seasonally-aggregated values (one value per month per gridcell)
\item comparisons: Output of \link[DGVMTools:fullSpatialComparison]{DGVMTools::fullSpatialComparison}
\item tables: List with two named data frames: "totals" and "metrics"
\item benchmark: A \link[DGVMBenchmarks:DaveBenchmark]{DGVMBenchmarks::DaveBenchmark} instance
}}

\item{settings}{Benchmarking settings as specified by the user}

\item{params}{Benchmarking parameters as specified by the user}

\item{overall_tables}{List of all tables}

\item{title}{Title for the plots}
}
\value{
Returns a list of HTML tags for rendering, or NULL if DGVMBenchmarks is not available
}
\description{
Produce a standard set of plots for the tellaus report from the given
benchmark.
}
