% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-csv.r
\name{get_field_csv}
\alias{get_field_csv}
\title{Get a Field for CSV}
\usage{
get_field_csv(
  source,
  quant,
  layers = NULL,
  target.STAInfo,
  file.name,
  verbose = FALSE,
  lat_col = "Lat",
  lon_col = "Lon",
  site_col = NULL,
  time_col = "date",
  sites = NULL
)
}
\arguments{
\item{source}{A \linkS4class{Source} containing the meta-data about the CSV source}

\item{quant}{A Quantity object to specify what quantity should be opened.}

\item{layers}{A character string (or a vector of character strings) specifying which variables from the CSV file are to be read.
NULL (default) means read all.}

\item{file.name}{Character string holding the name of the file.  This can be left blank, in which case the file name is automatically generated}

\item{verbose}{A logical, set to true to give progress/debug information}

\item{lat_col}{The name of the column containing latitude data}

\item{lon_col}{The name of the column containing longitude data}

\item{site_col}{The name of the column containing site names. Must specify this OR lat_col and lon_col.}

\item{time_col}{The name of the column containing time data}

\item{sites}{A character vector of site names to read. If NULL, read all sites.}

\item{target.sta.info}{An STAInfo object defining the spatial-temporal-annual extent over which we want the data}
}
\value{
A list containing firstly the data.table containing the data, and secondly the STAInfo for the data that we have
}
\description{
An internal function that reads data from a CSV .csv file.
}
\keyword{internal}
