% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozflux-benchmarks.r
\name{get_vars}
\alias{get_vars}
\title{Get the quantities plotted in the ozflux benchmarks.}
\usage{
get_vars()
}
\value{
Returns a list of \link[DGVMTools:Quantity]{DGVMTools::Quantity} objects.
}
\description{
Get the quantities plotted in the ozflux benchmarks.
}
\keyword{internal}
