% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tellaus.r
\name{make_metric_table}
\alias{make_metric_table}
\title{Create a table of metrics.}
\usage{
make_metric_table(benchmark, comparisons, simulations, ndigit = 3)
}
\arguments{
\item{benchmark}{A DGVMTools::Benchmark object containing metadata.}

\item{simulations}{A list of DGVMTools::Source objects.}

\item{ndigit}{The number of decimal digits in the output table.}

\item{comparison}{A list of DGVMTools::Comparison objects, typically the
output of fullSpatialComparison()[["Values"]], or the output of sequential
calls to CompareLayers(), concatenated into a single list.}
}
\value{
A data table.
}
\description{
The table will have the following columns:
- Dataset
- Quantity
- Simulation
}
\details{
The table will have one column per metric and one row per simulation.
}
\author{
Drew Holzworth \email{d.holzworth@westernsydney.edu.au}
}
\keyword{internal}
