% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tellaus.r
\name{make_metric_table}
\alias{make_metric_table}
\title{Create a table of metrics.}
\usage{
make_metric_table(benchmark, comparisons, simulations, ndigit = 3)
}
\arguments{
\item{benchmark}{A \link[DGVMTools:Benchmark]{DGVMTools::Benchmark} object containing metadata.}

\item{comparisons}{A list of \link[DGVMTools:Comparison]{DGVMTools::Comparison} objects, typically the
output of fullSpatialComparison()[\link{"Values"}], or the output of sequential
calls to CompareLayers(), concatenated into a single list.}

\item{simulations}{A list of \link[DGVMTools:Source]{DGVMTools::Source} objects.}

\item{ndigit}{The number of decimal digits in the output table.}
}
\value{
\link{data.frame} A data table.
}
\description{
The table will have the following columns:
\itemize{
\item Dataset
\item Quantity
\item Simulation
}
}
\details{
The table will have one column per metric and one row per simulation.
}
\keyword{internal}
