% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozflux-benchmarks.r
\name{ozflux_benchmarks}
\alias{ozflux_benchmarks}
\title{Generate ozflux benchmarks.}
\usage{
ozflux_benchmarks(
  sources,
  source_descriptions,
  use_plotly = FALSE,
  text_multiplier = 1.5,
  combined_graph = TRUE,
  max_title = 4,
  marker_size = 1,
  sites = NULL,
  vars = NULL,
  widget_dir = "plots",
  aspect_ratio = 16/9,
  show_all_observations = TRUE,
  show_all_predictions = TRUE
)
}
\arguments{
\item{use_plotly:}{Iff true, plots will be rendered using plotly. Note that
plotly plots are only interactive when rendering to HTML; therefore, we only
turn on plotly when rendering to HTML.This can of course be overriden (e.g.
for debugging).}

\item{combined_graph:}{Iff true, plots will be rendered in 2x2 panels for
each variable for each site.}

\item{max_title:}{Maximum number of nested titles.}

\item{marker_size:}{Marker size. Decrease for smaller markers.}

\item{sites:}{List of names of ozflux sites to be plotted. If NULL, all ozflux sites will be processed.}

\item{vars:}{List of names of output variables to be plotted. If NULL, a standard set of outputs will be plotted. This should be, for example, \code{c("dave_lai", "dave_resp")}.}

\item{widget_dir:}{Directory to which plotly plots will be saved.}

\item{aspect_ratio:}{Aspect ratio (width:height) used for plotly plots.}

\item{show_all_observations:}{If true, all observations will be returned (ie the predicted data may contain NA values). If false, only the observations which have a matching prediction will be returned.}

\item{show_all_predictions:}{If true, all predictions will be returned (ie the observed data may contain NA values). If false, only the predictions which have a matching observation will be returned.}
}
\value{
Returns a list of HTML tags if use_plotly is TRUE. Otherwise all
content will be printed, for use in R markdown.
}
\description{
Generate ozflux benchmarks.
}
