% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozflux-plotting.r
\name{ozflux_panel}
\alias{ozflux_panel}
\title{Create a 2x2 panel of plots, identical to what is seen in the ozflux
benchmarks.}
\usage{
ozflux_panel(
  sources,
  var,
  sites = NULL,
  use_plotly = FALSE,
  common_yaxis = FALSE
)
}
\arguments{
\item{sources:}{One or more sources of data. These may be paths to LPJ-GUESS
repositories, paths to directories containing output files, or as
\seealso{\link{DGVMTools::Source}} objects.}

\item{sites:}{One or more ozflux sites. May be specified as names of sites,
or as tuples of (Lon, Lat). If NULL, all sites will be plotted.}

\item{use_plotly:}{TRUE to generate plotly outputs, FALSE to use ggplot.}

\item{common_yaxis:}{TRUE to use the same y-axis scales for all sites. FALSE
otherwise.}
}
\value{
Returns a single plot object if plotting one site, otherwise returns
a list of plot objects of the same length as the number of sites to be
plotted.
}
\description{
One series will be plotted on each graph for each source. One
panel will be created for each site.
}
\author{
Drew Holzworth
}
