% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozflux-plotting.r
\name{ozflux_plot}
\alias{ozflux_plot}
\title{Create timeseries plots of predictions against observations.}
\usage{
ozflux_plot(
  sources,
  vars,
  sites = NULL,
  separate = FALSE,
  use_plotly = FALSE,
  common_yaxis = FALSE
)
}
\arguments{
\item{sources}{Sources to be plotted (see: \link{sanitise_source})}

\item{vars}{Variables to be plotted (see: \link{sanitise_variable})}

\item{sites}{Sites to be plotted (see: \link{sanitise_ozflux_sites})}

\item{separate}{iff TRUE, a list of the site-level plots will be returned.
If FALSE, all plots will be rendered to a single panel. This has no effect
if a single site is to be plotted.}

\item{use_plotly}{iff TRUE, graphs will be plotted using \link{plotly}}

\item{common_yaxis}{iff TRUE, all plots in the panel will have the same y-axis
range. No effect if only 1 site is being plotted.}
}
\value{
Returns either a single \link[ggplot2:ggplot]{ggplot2::ggplot} object (if separate is FALSE),
or a list of \link[ggplot2:ggplot]{ggplot2::ggplot} objects (if separate is TRUE). If use_plotly
is TRUE, returns \link{plotly} objects instead.
}
\description{
Plots a single variable across predictions at each specified ozflux site,
against observations (if any are available), generating one plot per site. If
separate is TRUE, then a separate plot will be returned for each site. If
separate is FALSE, then a panel of all plots will be returned.
}
\details{
If data is missing for any of the sites to be plotted, an error will occur.
}
