% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozflux-plotting.r
\name{ozflux_plot_layerwise}
\alias{ozflux_plot_layerwise}
\title{Create ozflux site-level layerwise plots.}
\usage{
ozflux_plot_layerwise(
  sources,
  var,
  sites = NULL,
  layers = NULL,
  title = NULL,
  separate_plots = TRUE,
  combine_sites = FALSE,
  use_plotly = FALSE,
  common_yaxis = FALSE
)
}
\arguments{
\item{sources:}{Data sources to be plotted (see: \link{sanitise_source}).}

\item{vars:}{The single variable to be plotted (see: \link{sanitise_variable}).}

\item{sites:}{Sites to be plotted (see: \link{sanitise_ozflux_sites}).}

\item{layers:}{The layers to be plotted. E.g. \code{paste0("sw_", 0:14)}.}

\item{title:}{The desired panel titles.}

\item{separate_plots:}{TRUE to draw each layer in a separate plot. FALSE to
draw all layers on the same plot.}

\item{combine_sites:}{TRUE to combine plots for each site into a single
panel. Requires separate_plots = FALSE.}

\item{use_plotly:}{TRUE to draw plots with plotly. FALSE to use ggplot2.}

\item{common_yaxis:}{TRUE to use a common y-axis for all plots. False otherwise.}
}
\value{
Returns a list of ggplot objects.
}
\description{
Create a panel of plots, where each plot in the panel contains one of the
layers of the specified variable. One panel will be generated for each site.
}
\details{
(In this nomenclature, a "layer" technically refers to a column in the output
file).
}
