% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozflux-plotting.r
\name{ozflux_plot_layerwise}
\alias{ozflux_plot_layerwise}
\title{Create ozflux site-level layerwise plots.}
\usage{
ozflux_plot_layerwise(
  sources,
  var,
  sites = NULL,
  layers = NULL,
  title = NULL,
  separate_plots = TRUE,
  combine_sites = FALSE,
  use_plotly = FALSE,
  common_yaxis = FALSE,
  show_all_observations = TRUE,
  show_all_predictions = TRUE
)
}
\arguments{
\item{sources}{Data sources to be plotted (see: \link{sanitise_source})}

\item{sites}{Sites to be plotted (see: \link{sanitise_ozflux_sites})}

\item{layers}{The layers to be plotted. E.g. \code{paste0("sw_", 0:14)}}

\item{title}{The desired panel titles}

\item{separate_plots}{iff TRUE, draw each layer in a separate plot. FALSE to
draw all layers in a single plot}

\item{use_plotly}{iff TRUE, generate \link{plotly} outputs, FALSE to use \link[ggplot2:ggplot]{ggplot2::ggplot}}

\item{common_yaxis}{iff TRUE, use a common y-axis for all plots. False otherwise}

\item{vars}{The single variable to be plotted (see: \link{sanitise_variable})}

\item{show_all_observations:}{If true, all observations will be returned (ie the predicted data may contain NA values). If false, only the observations which have a matching prediction will be returned.}

\item{show_all_predictions:}{If true, all predictions will be returned (ie the observed data may contain NA values). If false, only the predictions which have a matching observation will be returned.}
}
\value{
Returns a list of \link[ggplot2:ggplot]{ggplot2::ggplot} objects. If use_plotly is TRUE,
returns \link{plotly} objects instead.
}
\description{
Create a panel of plots, where each plot in the panel contains one of the
layers of the specified variable. One panel will be generated for each site.
}
\details{
(In this nomenclature, a "layer" technically refers to a column in the output
file).
}
