% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozflux-shiny.r
\name{ozflux_shiny}
\alias{ozflux_shiny}
\title{Run a shiny app which generates interactive plots of site-level ozflux
simulations against predictions.}
\usage{
ozflux_shiny(sources)
}
\arguments{
\item{sources}{List of \link[DGVMTools:Source]{DGVMTools::Source} objects to plot}

\item{vars}{Variables to plot (see: \link{get_observed_vars})}

\item{sites}{Sites to plot (see: \link{read_ozflux_sites})}
}
\value{
Returns a \link[shiny:shinyApp]{shiny::shinyApp} object
}
\description{
Run a shiny app which generates interactive plots of site-level ozflux
simulations against predictions.
}
