% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozflux-shiny.r
\name{ozflux_shiny}
\alias{ozflux_shiny}
\title{Run a shiny app which generates interactive plots of site-level ozflux
simulations against predictions.}
\usage{
ozflux_shiny(sources)
}
\description{
Run a shiny app which generates interactive plots of site-level ozflux
simulations against predictions.
}
\author{
Drew Holzworth
}
