% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tellaus.r
\name{plot_absolute_deltas}
\alias{plot_absolute_deltas}
\title{Create an absolute values differenes plot panel for use in the tellaus
benchmarks.}
\usage{
plot_absolute_deltas(benchmark, settings)
}
\arguments{
\item{benchmark}{Output of one of the benchmark_* functions.}

\item{settings}{Benchmark settings as defined in tellaus.rmd.}
}
\value{
\link{ggplot} A list of tags
}
\description{
Create an absolute values differenes plot panel for use in the tellaus
benchmarks.
}
