% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-auseflux.r
\name{read_auseflux}
\alias{read_auseflux}
\title{Read BoM LAI from the specified data file, or from the default location if
no data file is provided.}
\usage{
read_auseflux(var)
}
\arguments{
\item{var}{The AusEFlux variable.}
}
\value{
A DGVMTools::Field object containing the requested AusEFlux data.
}
\description{
Read BoM LAI from the specified data file, or from the default location if
no data file is provided.
}
\author{
Drew Holzworth \email{d.holzworth@westernsydney.edu.au}
}
