% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-lai.r
\name{read_bom_lai}
\alias{read_bom_lai}
\title{Read BoM LAI from the specified data file, or from the default location if
no data file is provided.}
\usage{
read_bom_lai(data_path, type = "rec")
}
\arguments{
\item{data_path}{Path to the data directory}

\item{type}{The type of LAI to read.Should be either "rec", "per", or "tot"}
}
\value{
Returns a \link[DGVMTools:Field]{DGVMTools::Field} object containing the BoM LAI data
}
\description{
Read BoM LAI from the specified data file, or from the default location if
no data file is provided.
}
