% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dave-processing.r
\name{read_data}
\alias{read_data}
\title{Read data for the specified variable.}
\usage{
read_data(sources, vars, sites = NULL, layers = NULL, correct_leaps = FALSE)
}
\arguments{
\item{sources:}{The data sources from which to read (see: \link{sanitise_source}).}

\item{vars:}{The variables to be read (see: \link{sanitise_variable}).}

\item{sites:}{The ozflux sites for which data will be read (see: \link{sanitise_ozflux_sites}).}

\item{layers:}{Names of layers to be read.}

\item{correct_leaps:}{Never set this to TRUE unless you know what you're doing.}
}
\value{
A single \link[DGVMTools:Field]{DGVMTools::Field} with a layer of observations, and one layer
for each source specified in sources.
}
\description{
Read data for the specified variable.
}
\keyword{internal}
