% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-sanitisation.r
\name{sanitise_ozflux_site}
\alias{sanitise_ozflux_site}
\title{Sanitise a single ozflux site name.}
\usage{
sanitise_ozflux_site(site)
}
\arguments{
\item{site:}{Name of the site, or a (lon, lat) tuple.}
}
\value{
Returns the site as a named vector with (Lon, Lat, Name).
}
\description{
Sanitise a single ozflux site name.
}
\author{
Drew Holzworth
}
\keyword{internal}
