% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-sanitisation.r
\name{sanitise_ozflux_sites}
\alias{sanitise_ozflux_sites}
\title{Sanitise the a site specifier.}
\usage{
sanitise_ozflux_sites(sites)
}
\arguments{
\item{sites:}{One or more ozflux site identifiers. If NULL, all sites will be
used. The sites may be expressed as:
\itemize{
\item Site names
\item (lon, lat) tuples
\item A dataframe with (Lon, Lat, Name) columns, and one row per site
}}
}
\value{
Returns a dataframe with (Lon, Lat, Name) columns, and one row per
site.
}
\description{
Sanitise the a site specifier.
}
\keyword{internal}
