% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-sanitisation.r
\name{sanitise_ozflux_sites}
\alias{sanitise_ozflux_sites}
\title{Sanitise the a site specifier.}
\usage{
sanitise_ozflux_sites(sites)
}
\arguments{
\item{sites:}{One or more ozflux site identifiers. These may be expressed as
site names or (lon, lat) tuples.}
}
\value{
Returns a list of named vectors with (lon, lat, name) elements.
}
\description{
Sanitise the a site specifier.
}
\author{
Drew Holzworth \email{d.holzworth@westernsydney.edu.au}
}
\keyword{internal}
