% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-sanitisation.r
\name{sanitise_sources}
\alias{sanitise_sources}
\title{Sanitise version inputs for plotting. Returns a list of \link[DGVMTools:Source]{DGVMTools::Source}
objects.}
\usage{
sanitise_sources(sources)
}
\arguments{
\item{sources}{A list of sources parsable by \link{sanitise_source}.}
}
\value{
Returns a list of Sources.
}
\description{
This is an internal helper function and is not intended for external use.
}
\keyword{internal}
