% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-sanitisation.r
\name{sanitise_variables}
\alias{sanitise_variables}
\title{Sanitise a list of variables provided by the user.}
\usage{
sanitise_variables(vars)
}
\arguments{
\item{vars:}{List of variables provided by the user (see: \link{sanitise_variable}).}
}
\value{
Returns a list of \link[DGVMTools:Quantity]{DGVMTools::Quantity} objects.
}
\description{
Sanitise a list of variables provided by the user.
}
\keyword{internal}
